requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.12
click==7.1.2
Flask==1.1.2
greenlet==1.1.2
gunicorn==20.0.4
idna==3.3
IMDbPY==2021.4.18
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.8.0
MarkupSafe==2.0.1
requests==2.27.1
SQLAlchemy==1.4.35
urllib3==1.26.9
Werkzeug==1.0.1
wincertstore==0.2
